Crate jsonrpc_http_server

jsonrpc http server.

extern crate jsonrpc_core;
extern crate jsonrpc_http_server;

use jsonrpc_core::*;
use jsonrpc_http_server::*;

fn main() {
    let mut io = IoHandler::new();
    io.add_method("say_hello", |_: Params| {
        Ok(Value::String("hello".to_string()))
    });

    let _server = ServerBuilder::new(io).start_http(&"127.0.0.1:3030".parse().unwrap());
}

Reexports

pub extern crate hyper;

Modules

tokio_core
Future-powered I/O at the core of Tokio

Structs

Host
Host type

Origin
Request Origin

Response
Simple server response structure

Rpc
RPC Handler bundled with metadata extractor.

Server
jsonrpc http server instance

ServerBuilder
Convenient JSON-RPC HTTP Server builder.

ServerHandler
jsonrpc http request handler.

Enums

AccessControlAllowOrigin
Origins allowed to access

CorsHeader
CORS Header Result.

DomainsValidation
Specifies if domains should be validated.

Error
RPC Server startup error.

RequestMiddlewareAction
Action undertaken by a middleware.

Traits

MetaExtractor
Extracts metadata from the HTTP request.

RequestMiddleware
Allows to intercept request and handle it differently.

Functions

cors_header
Returns a CORS header that should be returned with that request.

is_host_allowed
Returns true if Host header in request matches a list of allowed hosts.

Type Definitions

ServerResult
Result of starting the Server.