Struct jsonrpc_http_server::ServerBuilder

pub struct ServerBuilder<M: Metadata = (), S: Middleware<M> = NoopMiddleware> { /* fields omitted */ }

Convenient JSON-RPC HTTP Server builder.

Methods

impl<M: Metadata + Default, S: Middleware<M>> ServerBuilder<M, S>

pub fn new<T>(handler: T) -> Self where
T: Into<MetaIoHandler<M, S>>,

Creates new ServerBuilder for given IoHandler.

By default:

Server is not sending any CORS headers.

Server is validating Host header.

impl<M: Metadata, S: Middleware<M>> ServerBuilder<M, S>

pub fn with_meta_extractor<T, E>(handler: T, extractor: E) -> Self where
T: Into<MetaIoHandler<M, S>>,
E: MetaExtractor<M>,

Creates new ServerBuilder for given IoHandler.

By default:

Server is not sending any CORS headers.

Server is validating Host header.

pub fn event_loop_remote(self, remote: Remote) -> Self

Utilize existing event loop remote to poll RPC results.

Applies only to 1 of the threads. Other threads will spawn their own Event Loops.

pub fn rest_api(self, rest_api: RestApi) -> Self

Enable the REST -> RPC converter.

Allows you to invoke RPCs by sending POST /<method>/<param1>/<param2> requests (with no body). Disabled by default.

pub fn keep_alive(self, val: bool) -> Self

Sets Enables or disables HTTP keep-alive.

Default is true.

pub fn threads(self, threads: usize) -> Self

Sets number of threads of the server to run.

Panics when set to 0.

pub fn cors(
self,
cors_domains: DomainsValidation<AccessControlAllowOrigin>
) -> Self

Configures a list of allowed CORS origins.

pub fn cors_max_age<T: Into<Option<u32>>>(self, cors_max_age: T) -> Self

Configure CORS AccessControlMaxAge header returned.

Passing Some(millis) informs the client that the CORS preflight request is not necessary for at list millis ms. Disabled by default.

pub fn request_middleware<T: RequestMiddleware>(self, middleware: T) -> Self

Configures request middleware

pub fn meta_extractor<T: MetaExtractor<M>>(self, extractor: T) -> Self

Configures metadata extractor

pub fn allow_only_bind_host(self) -> Self

Allow connections only with Host header set to binding address.

pub fn allowed_hosts(self, allowed_hosts: DomainsValidation<Host>) -> Self

Specify a list of valid Host headers. Binding address is allowed automatically.

pub fn max_request_body_size(self, val: usize) -> Self

Sets the maximum size of a request body in bytes (default is 5 MiB).

pub fn start_http(self, addr: &SocketAddr) -> Result<Server>

Start this JSON-RPC HTTP server trying to bind to specified SocketAddr.